Seperated API and formatters into their own file
