+++ index.html
@@ -152,7 +152,8 @@
 
 
     <!-- Bootstrap.js -->
-    <script src="js/index.js"></script>
+    <script type="module" src="js/api.js"></script>
+    <script type="module" src="js/index.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.8/dist/js/bootstrap.bundle.min.js" integrity="sha384-FKyoEForCGlyvwx9Hj09JcYn3nv7wiPVlz7YYwJrWVcXK/BmnVDxM+D2scQbITxI" crossorigin="anonymous"></script>
 </body>
 </html>
