+++ pages/weather.html
@@ -34,6 +34,7 @@
         </div>
     </div>
     
+    <script src="../js/api.js"></script>
     <script src="../js/weather.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.8/dist/js/bootstrap.bundle.min.js" integrity="sha384-FKyoEForCGlyvwx9Hj09JcYn3nv7wiPVlz7YYwJrWVcXK/BmnVDxM+D2scQbITxI" crossorigin="anonymous"></script>
 </body>
